moved type function tests to the right package
