+++ src/test/java/graphql/annotations/typeFunctions/IntegerFunctionTests.java
@@ -12,16 +12,15 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  */
-package typeFunctions;
+package graphql.annotations.typeFunctions;
 
 import graphql.annotations.GraphQLAnnotations;
-import graphql.annotations.typeFunctions.DefaultTypeFunction;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import static graphql.Scalars.GraphQLInt;
 import static org.testng.Assert.assertEquals;
-import static typeFunctions.DefaultTypeFunctionTestHelper.testedDefaultTypeFunction;
+import static graphql.annotations.typeFunctions.DefaultTypeFunctionTestHelper.testedDefaultTypeFunction;
 
 public class IntegerFunctionTests {
