+++ src/test/java/graphql/annotations/typeFunctions/StringFunctionTests.java
@@ -12,18 +12,17 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  */
-package typeFunctions;
+package graphql.annotations.typeFunctions;
 
 import graphql.annotations.GraphQLAnnotations;
-import graphql.annotations.typeFunctions.DefaultTypeFunction;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import static graphql.Scalars.GraphQLBoolean;
+import static graphql.Scalars.GraphQLString;
 import static org.testng.Assert.assertEquals;
-import static typeFunctions.DefaultTypeFunctionTestHelper.testedDefaultTypeFunction;
+import static graphql.annotations.typeFunctions.DefaultTypeFunctionTestHelper.testedDefaultTypeFunction;
 
-public class BooleanFunctionTests {
+public class StringFunctionTests {
 
     @BeforeMethod
     public void init() {
@@ -31,9 +30,10 @@ public class BooleanFunctionTests {
     }
 
     @Test
-    public void buildType_booleanType_returnsGraphQLBoolean() {
+    public void buildType_stringType_returnsGraphQLString() {
         DefaultTypeFunction instance = testedDefaultTypeFunction();
-        assertEquals(instance.buildType(boolean.class, null,null), GraphQLBoolean);
-        assertEquals(instance.buildType(Boolean.class, null,null), GraphQLBoolean);
+        assertEquals(instance.buildType(String.class, null,null), GraphQLString);
     }
+
+
 }
